planner-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.34.0
firebase-admin==6.5.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pytz==2024.1
pydantic==2.9.0
